+++ integration-tests/fixtures/generate-bundle-stats/nextjs/nextjs-plugin.test.ts
@@ -82,6 +82,7 @@ describe("Generating nextjs stats", () => {
               entry: expect.any(Boolean),
               files: expect.arrayContaining([expect.any(String)]),
               names: expect.arrayContaining([expect.any(String)]),
+              dynamicImports: expect.any(Array),
             },
           ]),
           modules: expect.arrayContaining([
@@ -130,6 +131,7 @@ describe("Generating nextjs stats", () => {
               entry: expect.any(Boolean),
               files: expect.arrayContaining([expect.any(String)]),
               names: expect.arrayContaining([expect.any(String)]),
+              dynamicImports: expect.any(Array),
             },
           ]),
           modules: expect.arrayContaining([