+++ integration-tests/fixtures/generate-bundle-stats/nuxt/nuxt-plugin.test.ts
@@ -91,6 +91,7 @@ describe("Generating nuxt stats", () => {
                 entry: expect.any(Boolean),
                 files: expect.arrayContaining([expect.any(String)]),
                 names: expect.arrayContaining([expect.any(String)]),
+                dynamicImports: expect.any(Array),
               },
             ]),
             modules: expect.arrayContaining([
@@ -135,6 +136,7 @@ describe("Generating nuxt stats", () => {
                 entry: expect.any(Boolean),
                 files: expect.arrayContaining([expect.any(String)]),
                 names: expect.arrayContaining([expect.any(String)]),
+                dynamicImports: expect.any(Array),
               },
             ]),
             modules: expect.arrayContaining([