+++ integration-tests/fixtures/generate-bundle-stats/remix/remix-plugin.test.ts
@@ -88,6 +88,7 @@ describe("Generating remix stats", () => {
                 entry: expect.any(Boolean),
                 files: expect.arrayContaining([expect.any(String)]),
                 names: expect.arrayContaining([expect.any(String)]),
+                dynamicImports: expect.any(Array),
               },
             ]),
             modules: expect.arrayContaining([
@@ -130,6 +131,7 @@ describe("Generating remix stats", () => {
                 entry: expect.any(Boolean),
                 files: expect.arrayContaining([expect.any(String)]),
                 names: expect.arrayContaining([expect.any(String)]),
+                dynamicImports: expect.any(Array),
               },
             ]),
             modules: expect.arrayContaining([