+++ packages/bundle-analyzer/vitest.config.ts
@@ -1,6 +1,6 @@
 import replace from "@rollup/plugin-replace";
 import { defineProject } from "vitest/config";
-import { config, plugins } from "../../vitest.shared";
+import { config } from "../../vitest.shared";
 import "ts-node/register";
 
 // This file contains vitest configuration for testing the bundle-analyzer
@@ -32,6 +32,5 @@ export default defineProject({
       }),
       enforce: "pre",
     },
-    ...plugins,
   ],
 });