+++ packages/rollup-plugin/vitest.config.ts
@@ -1,13 +1,12 @@
 import replace from "@rollup/plugin-replace";
 import { defineProject } from "vitest/config";
-import { config, plugins } from "../../vitest.shared";
+import { config } from "../../vitest.shared";
 
 const packageJson = await import("./package.json");
 
 export default defineProject({
   ...config,
   plugins: [
-    // @ts-expect-error - using rollup plugin
     {
       ...replace({
         preventAssignment: true,
@@ -18,6 +17,5 @@ export default defineProject({
       }),
       enforce: "pre",
     },
-    ...plugins,
   ],
 });