+++ packages/webpack-plugin/src/webpack-bundle-analysis/webpackBundleAnalysisPlugin.ts
@@ -36,7 +36,7 @@ export const webpackBundleAnalysisPlugin: BundleAnalysisUploadPlugin = ({
           stage: webpack.Compilation.PROCESS_ASSETS_STAGE_REPORT,
         },
         async () => {
-          output.setBundleName(output.bundleName);
+          output.setBundleName(output.originalBundleName);
           // Webpack base chunk format options: https://webpack.js.org/configuration/output/#outputchunkformat
           if (typeof compilation.outputOptions.chunkFormat === "string") {
             if (compilation.name && compilation.name !== "") {
@@ -78,6 +78,8 @@ export const webpackBundleAnalysisPlugin: BundleAnalysisUploadPlugin = ({
             output.assets = collectedAssets;
           }
 
+          // need to collect all possible chunk ids beforehand
+          // this collection is done in the processChunks function
           const chunkIdMap = new Map<number | string, string>();
           if (chunks) {
             output.chunks = processChunks({ chunks, chunkIdMap });